Changed last_found_index to initialize from 0. I can do this because the increment of last_found_index is now in follow_path instead, and each loop using last_found_index starts from last_found_index instead of last_found_index + 1, which ensures the current segment isn't skipped.

+++ src/Template/display.cpp
@@ -131,7 +131,7 @@ void simple_screen_task(){
 }
 
 void print_curvepoints(std::vector<CurvePoint> points){
-  for(int i = 0; i < (int)points.size(); i++){
+  for(int i = 0; i < points.size(); i++){
     print_point(points[i].point, i);
     // TODO: also print drive_voltage, heading_max_voltage, and follow_distance
     // TODO: add a way to edit drive_voltage, heading_max_voltage, and follow_distance for each point
